backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==1.16.0
ecdsa==0.19.1
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
motor==3.3.2
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.6.1
python-dotenv==1.2.1
python-jose==3.3.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
